+++ src/Examples/TextClassification.cs
@@ -123,7 +123,7 @@ namespace TorchSharp.Examples
 
                     var loss = criterion(predicted_labels, labels);
                     loss.backward();
-                    torch.nn.utils.clip_grad_norm_(model.parameters(), 0.5);
+                    torch.nn.utils.clip_grad_norm_(model.parameters().ToArray(), 0.5);
                     optimizer.step();
 
                     total_acc += (predicted_labels.argmax(1) == labels).sum().to(torch.CPU).item<long>();