+++ src/Native/LibTorchSharp/THSNN.h
@@ -320,6 +320,41 @@ EXPORT_API(Tensor)   THSNN_GRUCell_forward(const NNModule module, const Tensor i
 EXPORT_API(NNModule) THSNN_LSTMCell_ctor(const int64_t input_size, const int64_t hidden_size, const bool bias, NNAnyModule* outAsAnyModule);
 EXPORT_API(Tensor)   THSNN_LSTMCell_forward(const NNModule module, const Tensor input1, const Tensor h0, const Tensor c0, Tensor* c_n);
 
+EXPORT_API(Tensor) THSNN_RNN_bias_ih(const NNModule module, const int64_t idx);
+EXPORT_API(void) THSNN_RNN_set_bias_ih(const NNModule module, const Tensor bias, const int64_t idx);
+EXPORT_API(Tensor) THSNN_RNN_weight_ih(const NNModule module, const int64_t idx);
+EXPORT_API(void) THSNN_RNN_set_weight_ih(const NNModule module, const Tensor weight, const int64_t idx);
+EXPORT_API(Tensor) THSNN_RNN_bias_hh(const NNModule module, const int64_t idx);
+EXPORT_API(void) THSNN_RNN_set_bias_hh(const NNModule module, const Tensor bias, const int64_t idx);
+EXPORT_API(Tensor) THSNN_RNN_weight_hh(const NNModule module, const int64_t idx);
+EXPORT_API(void) THSNN_RNN_set_weight_hh(const NNModule module, const Tensor weight, const int64_t idx);
+
+EXPORT_API(Tensor) THSNN_RNNCell_bias_ih(const NNModule module);
+EXPORT_API(void) THSNN_RNNCell_set_bias_ih(const NNModule module, const Tensor bias);
+EXPORT_API(Tensor) THSNN_RNNCell_weight_ih(const NNModule module);
+EXPORT_API(void) THSNN_RNNCell_set_weight_ih(const NNModule module, const Tensor weight);
+EXPORT_API(Tensor) THSNN_RNNCell_bias_hh(const NNModule module);
+EXPORT_API(void) THSNN_RNNCell_set_bias_hh(const NNModule module, const Tensor bias);
+EXPORT_API(Tensor) THSNN_RNNCell_weight_hh(const NNModule module);
+EXPORT_API(void) THSNN_RNNCell_set_weight_hh(const NNModule module, const Tensor weight);
+
+EXPORT_API(Tensor) THSNN_LSTMCell_bias_ih(const NNModule module);
+EXPORT_API(void) THSNN_LSTMCell_set_bias_ih(const NNModule module, const Tensor bias);
+EXPORT_API(Tensor) THSNN_LSTMCell_weight_ih(const NNModule module);
+EXPORT_API(void) THSNN_LSTMCell_set_weight_ih(const NNModule module, const Tensor weight);
+EXPORT_API(Tensor) THSNN_LSTMCell_bias_hh(const NNModule module);
+EXPORT_API(void) THSNN_LSTMCell_set_bias_hh(const NNModule module, const Tensor bias);
+EXPORT_API(Tensor) THSNN_LSTMCell_weight_hh(const NNModule module);
+EXPORT_API(void) THSNN_LSTMCell_set_weight_hh(const NNModule module, const Tensor weight);
+
+EXPORT_API(Tensor) THSNN_GRUCell_bias_ih(const NNModule module);
+EXPORT_API(void) THSNN_GRUCell_set_bias_ih(const NNModule module, const Tensor bias);
+EXPORT_API(Tensor) THSNN_GRUCell_weight_ih(const NNModule module);
+EXPORT_API(void) THSNN_GRUCell_set_weight_ih(const NNModule module, const Tensor weight);
+EXPORT_API(Tensor) THSNN_GRUCell_bias_hh(const NNModule module);
+EXPORT_API(void) THSNN_GRUCell_set_bias_hh(const NNModule module, const Tensor bias);
+EXPORT_API(Tensor) THSNN_GRUCell_weight_hh(const NNModule module);
+EXPORT_API(void) THSNN_GRUCell_set_weight_hh(const NNModule module, const Tensor weight);
 
 // Containers