+++ src/TorchSharp/NN/Convolution/Conv1D.cs
@@ -53,6 +53,7 @@ namespace TorchSharp
                 set {
                     THSNN_Conv1d_set_bias(handle, (value is null ? IntPtr.Zero : value.Handle));
                     torch.CheckForErrors();
+                    ConditionallyRegisterParameter("bias", value);
                 }
             }
             [DllImport("LibTorchSharp")]
@@ -69,6 +70,7 @@ namespace TorchSharp
                 set {
                     THSNN_Conv1d_set_weight(handle, value.Handle);
                     torch.CheckForErrors();
+                    ConditionallyRegisterParameter("weight", value);
                 }
             }
         }