+++ src/TorchSharp/NN/Convolution/ConvTranspose3D.cs
@@ -38,6 +38,7 @@ namespace TorchSharp
                 set {
                     THSNN_ConvTranspose3d_set_bias(handle, (value is null ? IntPtr.Zero : value.Handle));
                     torch.CheckForErrors();
+                    ConditionallyRegisterParameter("bias", value);
                 }
             }
             [DllImport("LibTorchSharp")]
@@ -54,6 +55,7 @@ namespace TorchSharp
                 set {
                     THSNN_ConvTranspose3d_set_weight(handle, value.Handle);
                     torch.CheckForErrors();
+                    ConditionallyRegisterParameter("weight", value);
                 }
             }
         }