+++ src/TorchSharp/NN/Embedding.cs
@@ -38,6 +38,7 @@ namespace TorchSharp
                 set {
                     THSNN_Embedding_set_weight(handle, value.Handle);
                     torch.CheckForErrors();
+                    ConditionallyRegisterParameter("weight", value);
                 }
             }
         }