+++ src/TorchSharp/NN/Normalization/BatchNorm1D.cs
@@ -1,11 +1,14 @@
 // Copyright (c) .NET Foundation and Contributors.  All Rights Reserved.  See LICENSE in the project root for license information.
 using System;
 using System.Runtime.InteropServices;
+using System.Collections.Generic;
+
 using static TorchSharp.torch;
 
 #nullable enable
 namespace TorchSharp
 {
+    using System.Collections.Generic;
     using Modules;
 
     namespace Modules
@@ -54,6 +57,7 @@ namespace TorchSharp
                 set {
                     THSNN_BatchNorm1d_set_bias(handle, (value is null ? IntPtr.Zero : value.Handle));
                     torch.CheckForErrors();
+                    ConditionallyRegisterParameter("bias", value);
                 }
             }
 
@@ -66,6 +70,7 @@ namespace TorchSharp
                 set {
                     THSNN_BatchNorm1d_set_weight(handle, value.Handle);
                     torch.CheckForErrors();
+                    ConditionallyRegisterParameter("weight", value);
                 }
             }
 
@@ -83,6 +88,14 @@ namespace TorchSharp
                     if (res == IntPtr.Zero) { torch.CheckForErrors(); return null; }
                     return new Tensor(res);
                 }
+            }
+
+            protected override void state_dict(Dictionary<string, Tensor> res)
+            {
+                var v = running_var;
+                var m = running_mean;
+                if (v is not null) res.Add("running_var", v);
+                if (m is not null) res.Add("running_mean", m);
             }
 
             public void reset_running_stats()