+++ src/TorchSharp/NN/Recurrent/GRUCell.cs
@@ -6,7 +6,7 @@ using System.Runtime.InteropServices;
 using static TorchSharp.torch;
 using static TorchSharp.torch.nn;
 
-
+#nullable enable
 namespace TorchSharp
 {
     using Modules;
@@ -32,11 +32,81 @@ namespace TorchSharp
             /// <param name="input">Tensor of shape (batch, input_size) containing the features of the input sequence.</param>
             /// <param name="h0">Tensor of shape (batch, hidden_size) containing the initial hidden state for each element in the batch.</param>
             /// <returns></returns>
-            public override Tensor forward(Tensor input, Tensor h0 = null)
+            public override Tensor forward(Tensor input, Tensor? h0 = null)
             {
                 var hN = THSNN_GRUCell_forward(handle, input.Handle, h0?.Handle ?? IntPtr.Zero);
                 if (hN == IntPtr.Zero) { torch.CheckForErrors(); }
                 return new Tensor(hN);
+            }
+
+            [DllImport("LibTorchSharp")]
+            extern static IntPtr THSNN_GRUCell_bias_ih(torch.nn.Module.HType module);
+            [DllImport("LibTorchSharp")]
+            extern static void THSNN_GRUCell_set_bias_ih(torch.nn.Module.HType module, IntPtr tensor);
+            [DllImport("LibTorchSharp")]
+            extern static IntPtr THSNN_GRUCell_bias_hh(torch.nn.Module.HType module);
+            [DllImport("LibTorchSharp")]
+            extern static void THSNN_GRUCell_set_bias_hh(torch.nn.Module.HType module, IntPtr tensor);
+
+            public Tensor? bias_ih {
+                get {
+                    var res = THSNN_GRUCell_bias_ih(handle);
+                    if (res == IntPtr.Zero) { torch.CheckForErrors(); }
+                    return ((res == IntPtr.Zero) ? null : new Tensor(res));
+                }
+                set {
+                    THSNN_GRUCell_set_bias_ih(handle, (value is null ? IntPtr.Zero : value.Handle));
+                    torch.CheckForErrors();
+                    ConditionallyRegisterParameter("bias_ih", value);
+                }
+            }
+
+            public Tensor? bias_hh {
+                get {
+                    var res = THSNN_GRUCell_bias_hh(handle);
+                    if (res == IntPtr.Zero) { torch.CheckForErrors(); }
+                    return ((res == IntPtr.Zero) ? null : new Tensor(res));
+                }
+                set {
+                    THSNN_GRUCell_set_bias_hh(handle, (value is null ? IntPtr.Zero : value.Handle));
+                    torch.CheckForErrors();
+                    ConditionallyRegisterParameter("bias_hh", value);
+                }
+            }
+
+            [DllImport("LibTorchSharp")]
+            extern static IntPtr THSNN_GRUCell_weight_ih(torch.nn.Module.HType module);
+            [DllImport("LibTorchSharp")]
+            extern static void THSNN_GRUCell_set_weight_ih(torch.nn.Module.HType module, IntPtr tensor);
+            [DllImport("LibTorchSharp")]
+            extern static IntPtr THSNN_GRUCell_weight_hh(torch.nn.Module.HType module);
+            [DllImport("LibTorchSharp")]
+            extern static void THSNN_GRUCell_set_weight_hh(torch.nn.Module.HType module, IntPtr tensor);
+
+            public Tensor weight_ih {
+                get {
+                    var res = THSNN_GRUCell_weight_ih(handle);
+                    if (res == IntPtr.Zero) { torch.CheckForErrors(); }
+                    return new Tensor(res);
+                }
+                set {
+                    THSNN_GRUCell_set_weight_ih(handle, value.Handle);
+                    torch.CheckForErrors();
+                    ConditionallyRegisterParameter("weight_ih", value);
+                }
+            }
+
+            public Tensor weight_hh {
+                get {
+                    var res = THSNN_GRUCell_weight_hh(handle);
+                    if (res == IntPtr.Zero) { torch.CheckForErrors(); }
+                    return new Tensor(res);
+                }
+                set {
+                    THSNN_GRUCell_set_weight_hh(handle, value.Handle);
+                    torch.CheckForErrors();
+                    ConditionallyRegisterParameter("weight_hh", value);
+                }
             }
         }
     }